backend/requirements.txt
flask==3.1.1
flask-socketio==5.3.3
eventlet==0.40.1
kafka-python==2.0.2
dnspython==2.7.0
greenlet==3.2.3
